refactor: Money 패키지명 수정
