--- src/test/java/domain/fixture/MoneyFixture.java
@@ -1,14 +1,0 @@
-package domain.fixture;
-
-import domain.bank.Money;
-
-/**
- * @author 우가
- * @version 1.0.0
- * @Created by 우가 on 2023/03/11
- */
-public class MoneyFixture {
-
-    public static final Money 만원 = new Money(10000);
-    public static final Money 이만원 = new Money(20000);
-}
